feat : HomeScreen 날씨 API 연동 완료

+++ domain/weather/src/main/java/com/umcspot/spot/weather/repository/WeatherRepository.kt
@@ -3,5 +3,5 @@ package com.umcspot.spot.weather.repository
 import com.umcspot.spot.weather.model.WeatherResult
 
 interface WeatherRepository {
-    suspend fun getWeather(longitude : Long, latitude : Long): Result<WeatherResult>
+    suspend fun getWeather(longitude : Double, latitude : Double): Result<WeatherResult>
 }
